+++ class/commondocgeneratorreferenceletters.class.php
@@ -89,6 +89,9 @@ class CommonDocGeneratorReferenceLetters extends CommonDocGenerator
 				$resarray['object_references'] = '';
 			}
 		}
+		if ($object->element == 'commandeFournisseur'){
+
+		}
 		// contact emetteur
 		$arrayidcontact = $object->getIdContact('internal', 'SALESREPFOLL');
 		$resarray[$array_key . '_contactsale'] = '';
@@ -1193,8 +1196,7 @@ class CommonDocGeneratorReferenceLetters extends CommonDocGenerator
 						// Attention, ce test est différent d'un isset()
 						if (is_array($object->array_options) && count($object->array_options)>0 && array_key_exists('options_'.$key_opt, $object->array_options))
 						{
-							$val = $this->showOutputFieldValue($extrafields, $key_opt, $object->array_options['options_'.$key_opt]);
-
+							$val = $this->showOutputFieldValue($extrafields, $key_opt, $object->array_options['options_'.$key_opt],'', $object->table_element);
 							$array_other['object_options_'.$key_opt] = $val;
 							$array_other['object_array_options_options_'.$key_opt] = $val;
 						}
@@ -1301,7 +1303,7 @@ class CommonDocGeneratorReferenceLetters extends CommonDocGenerator
 			    $hidden= ($ishidden == 0 ?  1 : 0);
 			}
 			else{
-			    $hidden=(($list == 0) ? 1 : 0);		// If zero, we are sure it is hidden, otherwise we show. If it depends on mode (view/create/edit form or list, this must be filtered by caller)
+			    $hidden=(($list == 0)  ? 1 : 0);		// If zero, we are sure it is hidden, otherwise we show. If it depends on mode (view/create/edit form or list, this must be filtered by caller)
 			}
 		}
 		if ($hidden) return '';		// This is a protection. If field is hidden, we should just not call this method.
@@ -1610,8 +1612,6 @@ class CommonDocGeneratorReferenceLetters extends CommonDocGenerator
 				$extrafields->attribute_hidden = $extrafields->attributes[$object->table_element]['hidden'] ?? array();
 			}
 		}
-
-
 		// phpcs:enable
 		global $conf;
 		if (!empty($extrafields->attribute_label && is_array($extrafields->attribute_label))){
@@ -1707,6 +1707,18 @@ class CommonDocGeneratorReferenceLetters extends CommonDocGenerator
 		}
 
 
+		// POUR LE MOMENT ON N'AJOUTE QUE LSE EXTRAFIELDS DE COMMANDE FOURNISSEURS
+		echo '<pre>' . var_export(get_class($object), true) . '</pre>';
+		$TallowedExtraInModeles = array('CommandeFournisseur','Facture');
+		if ( in_array(get_class($object), $TallowedExtraInModeles) ){
+			// Ajout des extrafields dans la selection des substitutions
+			if (!empty($extrafields->attributes[$object->table_element]['label'])) {
+				foreach ($extrafields->attributes[$object->table_element]['label'] as $key => $val) {
+					//Add value to store
+					$array_to_fill=array_merge($array_to_fill, array($array_key.'_options_'.$key => $object->array_options['options_'.$key]));
+				}
+			}
+		}
 		return $array_to_fill;
 	}
 }